updated for transformer install

+++ main.py
@@ -1,5 +1,6 @@
 import configparser
 import os
+import shutil
 import stat
 import time
 import warnings
@@ -10,12 +11,14 @@ from streamsets.sdk import ControlHub
 
 start_time = time.time()
 # sys.path.insert(1, os.path.abspath('/Users/sanjeev/SDK_4x'))
+INSTALLATION_HOME = os.getenv("HOME")
 config = configparser.ConfigParser()
 config.optionxform = lambda option: option
-config.read('credentials.properties')
+config.read('private/credentials.properties')
 CRED_ID = config.get("SECURITY", "CRED_ID")
 CRED_TOKEN = config.get("SECURITY", "CRED_TOKEN")
 SLACK_WEBHOOK = config.get("SECURITY", "SLACK_WEBHOOK")
+GMAIL_CRED = config.get("SECURITY", "GMAIL_CRED")
 
 config.read('deployment.conf')
 ENVIRONMENT_NAME = config.get("DEPLOYMENT", "ENVIRONMENT_NAME")
@@ -26,6 +29,7 @@ DEPLOYMENT_TYPE = config.get("DEPLOYMENT", "DEPLOYMENT_TYPE")
 DEPLOYMENT_TAGS = config.get("DEPLOYMENT", "DEPLOYMENT_TAGS")
 ENGINE_TYPE = config.get("DEPLOYMENT", "ENGINE_TYPE")
 ENGINE_VERSION = config.get("DEPLOYMENT", "ENGINE_VERSION")
+SCALA_VERSION = config.get("DEPLOYMENT", "SCALA_VERSION")
 ENGINE_LABELS = config.get("DEPLOYMENT", "ENGINE_LABELS")
 INSTALL_TYPE = config.get("DEPLOYMENT", "INSTALL_TYPE")
 EXTERNAL_RESOURCES_PATH_DOCKER = config.get("DEPLOYMENT", "EXTERNAL_RESOURCES_PATH_DOCKER")
@@ -36,12 +40,10 @@ MIN_HEAP = config.get("DEPLOYMENT", "MIN_HEAP")
 EMAIL_ADDRESS = config.get("DEPLOYMENT", "EMAIL_ADDRESS")
 DOCKER_NETWORK = config.get("DEPLOYMENT", "DOCKER_NETWORK")
 DOCKER_PORTS = config.get("DEPLOYMENT", "DOCKER_PORTS")
-
 # Download MySql JDBC driver jar for demo pipeline
-mysql_jdbc_driver_url = "https://dev.mysql.com/get/Downloads/Connector-J/mysql-connector-java-8.0.23.tar.gz"
+MYSQL_JDBC_DRIVER_URL = config.get("MISC", "MYSQL_JDBC_DRIVER_URL")
 
-# Get environment variables
-GMAIL_CRED = os.environ.get('GMAIL_CRED')
+
 
 warnings.simplefilter("ignore")
 # Connect to the StreamSets DataOps Platform.
@@ -50,6 +52,7 @@ sch = ControlHub(credential_id=CRED_ID, token=CRED_TOKEN)
 
 def create_deployment():
     # Instantiate an EnvironmentBuilder instance to build an environment, and activate it.
+    #global engine_properties
     environment_builder = sch.get_environment_builder(environment_type=ENVIRONMENT_TYPE)
     environment = environment_builder.build(environment_name=ENVIRONMENT_NAME,
                                             environment_type=ENVIRONMENT_TYPE,
@@ -69,6 +72,15 @@ def create_deployment():
                                           engine_type=ENGINE_TYPE,
                                           engine_version=ENGINE_VERSION,
                                           deployment_tags=[f'{DEPLOYMENT_TAGS}'])
+    if ENGINE_TYPE == 'TF':
+        deployment = deployment_builder.build(deployment_name=DEPLOYMENT_NAME,
+                                              deployment_type=DEPLOYMENT_TYPE,
+                                              environment=environment,
+                                              engine_type=ENGINE_TYPE,
+                                              engine_version=ENGINE_VERSION,
+                                              scala_binary_version=SCALA_VERSION,
+                                              deployment_tags=[f'{DEPLOYMENT_TAGS}'])
+
     # Set engine labels
     labels = ENGINE_LABELS.split(",")
     deployment.engine_configuration.engine_labels = labels
@@ -83,48 +95,68 @@ def create_deployment():
     # retrieve deployment to make changes
     deployment = sch.deployments.get(deployment_name=DEPLOYMENT_NAME)
 
-    # Fewer stage libs for quick deployment
-    with open('partial_stage_libs.conf', 'r') as f:
-        for rec in f:
-            if rec.startswith('#'):
-                continue
-            deployment.engine_configuration.stage_libs.append(rec.rstrip())
-    deployment.engine_configuration.stage_libs = [f"{lib}:{current_engine_version}" for lib in
-                                                  deployment.engine_configuration.stage_libs]
 
-    # # Full list of stage libs for complete deployment with all stages
-    # with open('stage_libs.conf', 'r') as f:
-    #     for rec in f:
-    #         if rec.startswith('#'): continue
-    #         deployment.engine_configuration.stage_libs.append(rec.rstrip())
-    # deployment.engine_configuration.stage_libs = [f"{lib}:{current_engine_version}" for lib in
-    #                                               deployment.engine_configuration.stage_libs]
+    if ENGINE_TYPE == 'DC':
+        # Fewer stage libs for quick deployment
+        with open('config/sdc/sdc_partial_stage_libs.conf', 'r') as f:
+            for rec in f:
+                if rec.startswith('#'):
+                    continue
+                deployment.engine_configuration.stage_libs.append(rec.rstrip())
+        deployment.engine_configuration.stage_libs = [f"{lib}:{current_engine_version}" for lib in
+                                                      deployment.engine_configuration.stage_libs]
 
-    # get list of enterprise libs
-    # with open('enterprise_libs.conf', 'r') as f:
-    #     for rec in f:
-    #         if rec.startswith('#'):
-    #             continue
-    #         deployment.engine_configuration.stage_libs.append(rec.rstrip())
+        # # Full list of stage libs for complete deployment with all stages
+        # with open('sdc_stage_libs.conf', 'r') as f:
+        #     for rec in f:
+        #         if rec.startswith('#'): continue
+        #         deployment.engine_configuration.stage_libs.append(rec.rstrip())
+        # deployment.engine_configuration.stage_libs = [f"{lib}:{current_engine_version}" for lib in
+        #                                               deployment.engine_configuration.stage_libs]
 
-    # retrieve deployment configs
-    sdc_properties = javaproperties.loads(
-        deployment.engine_configuration.advanced_configuration.data_collector_configuration)
 
-    # read SDC properties
-    for key in config['SDC_PROPERTIES']:
-        sdc_properties[key] = config['SDC_PROPERTIES'][key]
+        # retrieve deployment configs
+        engine_properties = javaproperties.loads(
+            deployment.engine_configuration.advanced_configuration.data_collector_configuration)
+    if ENGINE_TYPE == 'TF':
+        engine_properties = javaproperties.loads(
+            deployment.engine_configuration.advanced_configuration.transformer_configuration)
+        # Fewer stage libs for quick deployment
+        with open('config/transformer/transformer_partial_stage_libs.conf', 'r') as f:
+            for rec in f:
+                if rec.startswith('#'):
+                    continue
+                deployment.engine_configuration.stage_libs.append(rec.rstrip())
+        deployment.engine_configuration.stage_libs = [f"{lib}:{current_engine_version}" for lib in
+                                                      deployment.engine_configuration.stage_libs]
+
+        # # Full list of stage libs for complete deployment with all stages
+        # with open('transformer_stage_libs.conf', 'r') as f:
+        #     for rec in f:
+        #         if rec.startswith('#'): continue
+        #         deployment.engine_configuration.stage_libs.append(rec.rstrip())
+        # deployment.engine_configuration.stage_libs = [f"{lib}:{current_engine_version}" for lib in
+        #                                               deployment.engine_configuration.stage_libs]
+
+        # retrieve deployment configs
+
+    # read engine properties
+    for key in config['ENGINE_PROPERTIES']:
+        engine_properties[key] = config['ENGINE_PROPERTIES'][key]
 
     # read runtime resources
     for key in config['RUNTIME_RESOURCES']:
-        sdc_properties[key] = config['RUNTIME_RESOURCES'][key]
+        engine_properties[key] = config['RUNTIME_RESOURCES'][key]
 
     if platform == "darwin":
         print("##### Mac Os Detected #####")
-        sdc_properties['sdc.base.http.url'] = 'http://localhost:18630'
+        engine_properties['sdc.base.http.url'] = 'http://localhost:18630'
 
-    properties = javaproperties.dumps(sdc_properties)
-    deployment.engine_configuration.advanced_configuration.data_collector_configuration = properties
+    engine_configurations = javaproperties.dumps(engine_properties)
+    if ENGINE_TYPE == 'DC':
+        deployment.engine_configuration.advanced_configuration.data_collector_configuration = engine_configurations
+    if ENGINE_TYPE == 'TF':
+        deployment.engine_configuration.advanced_configuration.transformer_configuration = engine_configurations
     # verify persisted config changes
     # deployment = sch.deployments.get(deployment_name=DEPLOYMENT_NAME)
     # print(javaproperties.loads(deployment.engine_configuration.advanced_configuration.data_collector_configuration)[
@@ -142,11 +174,12 @@ def create_deployment():
     java_config.java_memory_strategy = 'ABSOLUTE'
     java_config.maximum_java_heap_size_in_mb = MAX_HEAP
     java_config.minimum_java_heap_size_in_mb = MIN_HEAP
-    with open('java_options.conf', 'r') as f:
-        for rec in f:
-            if rec.startswith('#'):
-                continue
-            java_config.java_options = f"{java_config.java_options} {rec.rstrip()}"
+    if ENGINE_TYPE == 'DC':
+        with open('config/common/java_options.conf', 'r') as f:
+            for rec in f:
+                if rec.startswith('#'):
+                    continue
+                java_config.java_options = f"{java_config.java_options} {rec.rstrip()}"
 
     # configure aws credential store
     cred_properties = javaproperties.loads(deployment.engine_configuration.advanced_configuration.credential_stores)
@@ -156,8 +189,12 @@ def create_deployment():
     cred_properties = javaproperties.dumps(cred_properties)
     deployment.engine_configuration.advanced_configuration.credential_stores = cred_properties
 
-    properties = javaproperties.dumps(sdc_properties)
-    deployment.engine_configuration.advanced_configuration.data_collector_configuration = properties
+    engine_configurations = javaproperties.dumps(engine_properties)
+    if ENGINE_TYPE == 'DC':
+        deployment.engine_configuration.advanced_configuration.data_collector_configuration = engine_configurations
+    if ENGINE_TYPE == 'TF':
+        deployment.engine_configuration.advanced_configuration.transformer_configuration = engine_configurations
+
     # persist changes to the deployment
     sch.update_deployment(deployment)
     sch.start_deployment(deployment)
@@ -169,13 +206,13 @@ def create_deployment():
     update_config['DEPLOYMENT']['DEPLOYMENT_ID'] = str(deployment.deployment_id)
     update_config['DEPLOYMENT']['ENVIRONMENT_ID'] = str(environment.environment_id)
 
-    with open('deployment.conf', 'w') as configfile:  # save
+    with open('config/common/deployment.conf', 'w') as configfile:  # save
         update_config.write(configfile)
 
     if INSTALL_TYPE == "DOCKER":
         # engine version string to include in docker container name
-        if 'http.port' in config['SDC_PROPERTIES']:
-            engine_version = config['SDC_PROPERTIES']['http.port']
+        if 'http.port' in config['ENGINE_PROPERTIES']:
+            engine_version = config['ENGINE_PROPERTIES']['http.port']
         else:
             engine_version = current_engine_version.replace(".", "")
 
@@ -184,12 +221,20 @@ def create_deployment():
         ports_string = ' '.join(str(port) for port in ports_list)
         if not ports_list:
             # run SDC container under Docker 'cluster' network
-            install_script = deployment.install_script().replace("docker run",
+            if ENGINE_TYPE == 'DC':
+                install_script = deployment.install_script().replace("docker run",
                                                                  f"docker run --network=cluster  -h sdc.cluster --name sdc-{engine_version} ")
+            if ENGINE_TYPE == 'TF':
+                install_script = deployment.install_script().replace("docker run",
+                                                                 f"docker run --network=cluster  -h transformer.cluster --name tf-{engine_version} ")
         else:
             # run SDC container under Docker 'cluster' network +expose some Docker ports
-            install_script = deployment.install_script().replace("docker run",
+            if ENGINE_TYPE == 'DC':
+                install_script = deployment.install_script().replace("docker run",
                                                                  f"docker run --network=cluster  -h sdc.cluster --name sdc-{engine_version} {ports_string} ")
+            if ENGINE_TYPE == 'TF':
+                install_script = deployment.install_script().replace("docker run",
+                                                                     f"docker run --network=cluster  -h transformer.cluster --name tf-{engine_version} {ports_string} ")
 
         with open("install_script.sh", "w") as f:
             f.write('docker network create cluster\n')
@@ -197,10 +242,15 @@ def create_deployment():
         os.chmod("install_script.sh", stat.S_IRWXU)
         os.system("sh install_script.sh")
     if INSTALL_TYPE == "TARBALL":
-        install_script = deployment.install_script().replace("--foreground", "--background")
+        #install_script = deployment.install_script().replace("--foreground", "--background")
+        install_script = deployment.install_script(install_mechanism='BACKGROUND')
         # defaults the download & install location
-        install_script = f"{install_script} --no-prompt --download-dir=$HOME/.streamsets/download/dc " \
+        if ENGINE_TYPE == 'DC':
+            install_script = f"{install_script} --no-prompt --download-dir=$HOME/.streamsets/download/dc " \
                          f"--install-dir=$HOME/.streamsets/install/dc "
+        if ENGINE_TYPE == 'TF':
+            install_script = f"{install_script} --no-prompt --download-dir=$HOME/.streamsets/download/tf " \
+                         f"--install-dir=$HOME/.streamsets/install/tf "
         with open("install_script.sh", "w") as f:
             f.write('ulimit -n 32768\n')
             # if there's a requirement(for instance proxy configurations) to pass in some java options during
@@ -224,8 +274,8 @@ def delete_deployment():
 
         deployment = sch.deployments.get(deployment_id=deployment_id)
         current_engine_version = deployment.engine_configuration.engine_version
-        if 'http.port' in config['SDC_PROPERTIES']:
-            engine_version = config['SDC_PROPERTIES']['http.port']
+        if 'http.port' in config['ENGINE_PROPERTIES']:
+            engine_version = config['ENGINE_PROPERTIES']['http.port']
         else:
             engine_version = current_engine_version.replace(".", "")
         sch.delete_deployment(deployment)
@@ -242,19 +292,24 @@ def delete_deployment():
 
     try:
         if INSTALL_TYPE == "TARBALL":
+            installation_dir = f"{INSTALLATION_HOME}/.streamsets/install/dc/streamsets-datacollector-{ENGINE_VERSION}"
             with open("cleanup_script.sh", "w") as f:
                 f.write(
                     f"pid=`ps aux | grep streamsets-datacollector-{current_engine_version} | grep DataCollectorMain | awk {{'print $2'}}`\n")
                 f.write(f"kill -9 $pid\n")
+                f.write(f"rm -rf {installation_dir}\n")
                 f.write('echo "Finished cleanup tasks"\n')
         if INSTALL_TYPE == "DOCKER":
             with open("cleanup_script.sh", "w") as f:
-                f.write(f"docker rm -f sdc-{engine_version}\n")
+                if ENGINE_TYPE == 'TF':
+                    f.write(f"docker rm -f tf-{engine_version}\n")
+                if ENGINE_TYPE == 'DC':
+                    f.write(f"docker rm -f sdc-{engine_version}\n")
                 f.write('echo "Finished cleanup tasks"\n')
         os.chmod("cleanup_script.sh", stat.S_IRWXU)
         os.system("sh cleanup_script.sh")
     except:
-        print("DataCollector not running !!")
+        print("Engine not running !!")
     if os.path.exists("install_script.sh"):
         os.remove("install_script.sh")
     if os.path.exists("post_install_script.sh"):
@@ -277,7 +332,7 @@ def update_deployment():
 
     # Fewer stage libs for quick deployment
     deployment.engine_configuration.stage_libs = []
-    with open('partial_stage_libs.conf', 'r') as f:
+    with open('config/sdc/sdc_partial_stage_libs.conf', 'r') as f:
         for rec in f:
             if rec.startswith('#'):
                 continue
@@ -286,34 +341,28 @@ def update_deployment():
                                                   deployment.engine_configuration.stage_libs]
 
     # # Full list of stage libs for complete deployment with all stages
-    # with open('stage_libs.conf', 'r') as f:
+    # with open('sdc_stage_libs.conf', 'r') as f:
     #     for rec in f:
     #         if rec.startswith('#'): continue
     #         deployment.engine_configuration.stage_libs.append(rec.rstrip())
     # deployment.engine_configuration.stage_libs = [f"{lib}:{current_engine_version}" for lib in
     #                                               deployment.engine_configuration.stage_libs]
 
-    # get list of enterprise libs
-    with open('enterprise_libs.conf', 'r') as f:
-        for rec in f:
-            if rec.startswith('#'):
-                continue
-            deployment.engine_configuration.stage_libs.append(rec.rstrip())
 
     # retrieve deployment configs
-    sdc_properties = javaproperties.loads(
+    ENGINE_PROPERTIES = javaproperties.loads(
         deployment.engine_configuration.advanced_configuration.data_collector_configuration)
 
     # read SDC properties
-    for key in config['SDC_PROPERTIES']:
-        sdc_properties[key] = config['SDC_PROPERTIES'][key]
+    for key in config['ENGINE_PROPERTIES']:
+        ENGINE_PROPERTIES[key] = config['ENGINE_PROPERTIES'][key]
 
     # read runtime resources
     for key in config['RUNTIME_RESOURCES']:
-        sdc_properties[key] = config['RUNTIME_RESOURCES'][key]
+        ENGINE_PROPERTIES[key] = config['RUNTIME_RESOURCES'][key]
 
-    properties = javaproperties.dumps(sdc_properties)
-    deployment.engine_configuration.advanced_configuration.data_collector_configuration = properties
+    engine_configurations = javaproperties.dumps(ENGINE_PROPERTIES)
+    deployment.engine_configuration.advanced_configuration.data_collector_configuration = engine_configurations
 
     # Update external_resource_location
     if INSTALL_TYPE == "DOCKER":
@@ -328,7 +377,7 @@ def update_deployment():
     java_config.maximum_java_heap_size_in_mb = MIN_HEAP
     java_config.minimum_java_heap_size_in_mb = MAX_HEAP
     java_options = ""
-    with open('java_options.conf', 'r') as f:
+    with open('config/common/java_options.conf', 'r') as f:
         for rec in f:
             if rec.startswith('#'):
                 continue
@@ -343,8 +392,8 @@ def update_deployment():
     cred_properties = javaproperties.dumps(cred_properties)
     deployment.engine_configuration.advanced_configuration.credential_stores = cred_properties
 
-    properties = javaproperties.dumps(sdc_properties)
-    deployment.engine_configuration.advanced_configuration.data_collector_configuration = properties
+    engine_configurations = javaproperties.dumps(ENGINE_PROPERTIES)
+    deployment.engine_configuration.advanced_configuration.data_collector_configuration = engine_configurations
     # persist changes to the deployment
     sch.update_deployment(deployment)
     sch.stop_deployment(deployment)
